requirements-dev.txt
# Incluir requirements principales
-r requirements.txt

# Dependencias adicionales para desarrollo

# Testing y debugging
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2
django-debug-toolbar==4.2.0

# Formateo de código
black==23.7.0
flake8==6.0.0

# Documentación
sphinx==7.1.2

# Variables de entorno
python-decouple==3.8
